Changed the dataloader to not use the collate function, instead we can just tree map over the tensors to turn them into the format we need.
Tried out pickling the params to save after training and it worked.

Had to adjust the splitting to happen in the curricula loss calculation

Changed the dataset and dataloader to retunr the unsplit training
data points, Splitting now happens in the scan_batch.

The added scan function goes over the current number of batches in the epoch. Had to adjust it to feed saved data and not generate it on the fly.

Saving this now so I can work on it later on laptop, but its still very messy.