Cleaned up the SPL_simplified notebook.
Created and cleaned up aa_training_utils, like model creation code, dataset_generation code.

Created a training script that allows for providing a PRNG int seed to create a training for SPL with a given int seed dataset, so the two models baseline, and SPL can train on the same datasets for the 5 fold cross validation.  Could also be used for curricula grid searching.  Will just have to write a larger training script including this with a loop or something. NOTE: Might have to be careful to not keep previous runs in memory at all to overload the VRAM,