Added a parameter to the training  funciton for setting pytorch manual seed,
Added the evaluation functions, also added a periodical evaluation run with a certain dataset size using a sampler hard coded inside the training functions. created a single script to check on delftblue. TODO: Have to recheck if the Empirical cross entropy is calculated correctly, as collab showed it getting very negative. Not fully sure how to interpret that for now.

Also changed the fourier function creator to allow for period to be randomly sampled within a range. Same for the Shift , these could be used to create a more out of distribution set of tasks better.